Move container_init related function into new file
